Split gateway providers into three separate classes.
This commit partitions the mounting of endpoints into "gateway"-specific
classes. Each gateway provider includes routing information, controller registration
as a service (if applicable) and other gateway-specific configuration.
